Move Custom lockscreen wallpaper setting under lockscreen styles

+++ src/com/android/settings/AOSPAL/LockscreenStyle.java
@@ -18,16 +18,31 @@ package com.android.settings.AOSPAL;
 
 import android.app.Activity;
 import android.app.AlertDialog;
+import android.content.ActivityNotFoundException;
+import android.content.ComponentName;
 import android.app.Dialog;
 import android.app.DialogFragment;
+import android.database.ContentObserver;
+import android.content.ContentResolver;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.ServiceConnection;
+import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Point;
+import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.Environment;
+import android.os.Handler;
+import android.os.IBinder;
+import android.os.RemoteException;
+import android.os.UserHandle;
+import android.provider.Settings.SettingNotFoundException;
 import android.preference.CheckBoxPreference;
 import android.preference.ListPreference;
 import android.preference.Preference;
@@ -39,36 +54,41 @@ import android.provider.MediaStore;
 import android.provider.Settings;
 import android.util.Log;
 import android.util.TypedValue;
+import android.view.Display;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.Window;
+import android.view.View;
+import android.view.View.OnClickListener;
 import android.widget.Toast;
 
 import com.android.settings.SettingsPreferenceFragment;
 import com.android.settings.R;
 import com.android.settings.Utils;
+import com.android.internal.policy.IKeyguardService;
 
 import com.android.internal.widget.LockPatternUtils;
 
 import net.margaritov.preference.colorpicker.ColorPickerPreference;
 
 import java.io.File;
+import java.io.IOException;
 
 public class LockscreenStyle extends SettingsPreferenceFragment
         implements OnPreferenceChangeListener {
 
     private static final String TAG = "LockscreenStyle";
 
-    private static final String KEY_LOCKSCREEN_COLORIZE_ICON =
-            "lockscreen_colorize_icon";
-    private static final String KEY_LOCKSCREEN_LOCK_ICON =
-            "lockscreen_lock_icon";
-    private static final String KEY_LOCKSCREEN_FRAME_COLOR =
-            "lockscreen_frame_color";
-    private static final String KEY_LOCKSCREEN_LOCK_COLOR =
-            "lockscreen_lock_color";
-    private static final String KEY_LOCKSCREEN_DOTS_COLOR =
-            "lockscreen_dots_color";
+    private static final int REQUEST_CODE_BG_WALLPAPER = 1024;
+
+    private static final String KEY_LOCKSCREEN_COLORIZE_ICON = "lockscreen_colorize_icon";
+    private static final String KEY_LOCKSCREEN_LOCK_ICON = "lockscreen_lock_icon";
+    private static final String KEY_LOCKSCREEN_FRAME_COLOR = "lockscreen_frame_color";
+    private static final String KEY_LOCKSCREEN_LOCK_COLOR = "lockscreen_lock_color";
+    private static final String KEY_LOCKSCREEN_DOTS_COLOR = "lockscreen_dots_color";
+    private static final String KEY_LOCKSCREEN_WALLPAPER = "lockscreen_wallpaper";
+    private static final String KEY_SELECT_LOCKSCREEN_WALLPAPER = "select_lockscreen_wallpaper";
 
     private String mDefault;
 
@@ -79,6 +99,26 @@ public class LockscreenStyle extends SettingsPreferenceFragment
     private ColorPickerPreference mDotsColor;
 
     private ListPreference mLockIcon;
+
+    private CheckBoxPreference mLockscreenWallpaper;
+    private Preference mSelectLockscreenWallpaper;
+
+    private File mWallpaperTemporary;
+
+    private IKeyguardService mKeyguardService;
+
+    private final ServiceConnection mKeyguardConnection = new ServiceConnection() {
+        @Override
+        public void onServiceConnected(ComponentName name, IBinder service) {
+            mKeyguardService = IKeyguardService.Stub.asInterface(service);
+        }
+
+        @Override
+        public void onServiceDisconnected(ComponentName name) {
+            mKeyguardService = null;
+        }
+
+    };
 
     private boolean mCheckPreferences;
 
@@ -102,8 +142,16 @@ public class LockscreenStyle extends SettingsPreferenceFragment
             prefSet.removeAll();
         }
 
+        Intent intent = new Intent();
+        intent.setClassName("com.android.keyguard", "com.android.keyguard.KeyguardService");
+        if (!mContext.bindServiceAsUser(intent, mKeyguardConnection,
+                Context.BIND_AUTO_CREATE, UserHandle.OWNER)) {
+            Log.e(TAG, "*** Keyguard: can't bind to keyguard");
+        }
+
         addPreferencesFromResource(R.xml.lockscreen_style);
         prefSet = getPreferenceScreen();
+	ContentResolver resolver = getContentResolver();
 
         // Set to string so we don't have to create multiple objects of it
         mDefault = getResources().getString(R.string.default_string);
@@ -159,11 +207,20 @@ public class LockscreenStyle extends SettingsPreferenceFragment
         mLockIcon.setEnabled(!dotsDisabled);
         mColorizeCustom.setEnabled(!dotsDisabled && imageExists);
 
+        // Custom Lockscreen wallpaper
+        mLockscreenWallpaper = (CheckBoxPreference) findPreference(KEY_LOCKSCREEN_WALLPAPER);
+        mLockscreenWallpaper.setChecked(Settings.System.getInt(getContentResolver(), Settings.System.LOCKSCREEN_WALLPAPER, 0) == 1);
+
+        mSelectLockscreenWallpaper = findPreference(KEY_SELECT_LOCKSCREEN_WALLPAPER);
+        mSelectLockscreenWallpaper.setEnabled(mLockscreenWallpaper.isChecked());
+        mWallpaperTemporary = new File(getActivity().getCacheDir() + "/lockwallpaper.tmp");
+
         updateLockSummary();
 
         setHasOptionsMenu(true);
         mCheckPreferences = true;
         return prefSet;
+
     }
 
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
@@ -187,13 +244,81 @@ public class LockscreenStyle extends SettingsPreferenceFragment
                         Settings.Secure.LOCKSCREEN_LOCK_ICON, path);
 
                 mColorizeCustom.setEnabled(path != null);
+            } else if (requestCode == REQUEST_CODE_BG_WALLPAPER) {
+            if (resultCode == Activity.RESULT_OK) {
+                if (mWallpaperTemporary.length() == 0 || !mWallpaperTemporary.exists()) {
+                    Toast.makeText(getActivity(),
+                            getResources().getString(R.string.shortcut_image_not_valid),
+                            Toast.LENGTH_LONG).show();
+                    return;
+                }
+                Bitmap bmp = BitmapFactory.decodeFile(mWallpaperTemporary.getAbsolutePath());
+                try {
+                    mKeyguardService.setWallpaper(bmp);
+                    Settings.System.putInt(getContentResolver(), Settings.System.LOCKSCREEN_SEE_THROUGH, 0);
+                } catch (Exception ex) {
+                    Log.e(TAG, "Failed to set wallpaper: " + ex);
+                }
             }
+	}
+	if (mWallpaperTemporary.exists()) mWallpaperTemporary.delete();
         } else {
             if (mLockImage.exists()) {
                 mLockImage.delete();
             }
         }
         updateLockSummary();
+    }
+
+    @Override
+    public boolean onPreferenceTreeClick(PreferenceScreen preferenceScreen, Preference preference) {
+         ContentResolver cr = getActivity().getContentResolver();
+	 if (preference == mLockscreenWallpaper) {
+            if (!mLockscreenWallpaper.isChecked()) setWallpaper(null);
+            else Settings.System.putInt(getContentResolver(), Settings.System.LOCKSCREEN_WALLPAPER, 1);
+            mSelectLockscreenWallpaper.setEnabled(mLockscreenWallpaper.isChecked());
+        } else if (preference == mSelectLockscreenWallpaper) {
+            final Intent intent = new Intent(Intent.ACTION_PICK,
+                    android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
+            intent.setType("image/*");
+            intent.putExtra("crop", "true");
+            intent.putExtra("scale", true);
+            intent.putExtra("scaleUpIfNeeded", false);
+            intent.putExtra("outputFormat", Bitmap.CompressFormat.JPEG.toString());
+
+            final Display display = getActivity().getWindowManager().getDefaultDisplay();
+
+            boolean isPortrait = getResources().getConfiguration().orientation ==
+                    Configuration.ORIENTATION_PORTRAIT;
+
+            Point size = new Point();
+            display.getSize(size);
+
+            intent.putExtra("aspectX", isPortrait ? size.x : size.y);
+            intent.putExtra("aspectY", isPortrait ? size.y : size.x);
+
+            try {
+                mWallpaperTemporary.createNewFile();
+                mWallpaperTemporary.setWritable(true, false);
+                intent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(mWallpaperTemporary));
+                getActivity().startActivityFromFragment(this, intent, REQUEST_CODE_BG_WALLPAPER);
+            } catch (IOException e) {
+                // Do nothing here
+            } catch (ActivityNotFoundException e) {
+                // Do nothing here
+            }
+        }else {
+         return super.onPreferenceTreeClick(preferenceScreen, preference);
+       }
+         return true;
+    }
+
+    private void setWallpaper(Bitmap bmp) {
+        try {
+            mKeyguardService.setWallpaper(bmp);
+        } catch (RemoteException ex) {
+            Log.e(TAG, "Unable to set wallpaper!");
+        }
     }
 
     @Override
